feat(stop-advisory): rename from stop-momentum, externalize guidance [2.0.0]
BREAKING CHANGE: Plugin renamed from stop-momentum to stop-advisory. Config file renamed from .claude/momentum-guide.md to .claude/stop-guidance.md. Default guidance removed — hook is a no-op when unconfigured. Added STOP_HOOK_GUIDANCE env var as alternative config source.

Closes #150
