+++ plugins/stop-advisory/hooks/hooks.json
@@ -5,7 +5,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "uv run --script \"${CLAUDE_PLUGIN_ROOT}\"/hooks/stop-momentum.py"
+            "command": "uv run --script \"${CLAUDE_PLUGIN_ROOT}\"/hooks/stop-advisory.py"
           }
         ]
       }
