--- plugins/stop-momentum/CHANGELOG.md
@@ -1,9 +1,0 @@
-# Changelog
-
-## [1.0.0] - 2026-03-01
-
-### Added
-- Initial standalone release (split from `orchestration-discipline` v1.2.1)
-- Stop hook with ack-token handshake for deliberate session stops
-- Custom guidance support via `.claude/momentum-guide.md`
-- Per-session state storage with CLAUDE_HOOK_STATE_DIR override support
