+++ pyproject.toml
@@ -12,7 +12,7 @@ dev = [
 ]
 
 [tool.pytest.ini_options]
-testpaths = ["plugins/essentials/tests", "plugins/delegation-guard/tests", "plugins/stop-momentum/tests", "plugins/haiku-enforce/tests"]
+testpaths = ["plugins/essentials/tests", "plugins/delegation-guard/tests", "plugins/stop-advisory/tests", "plugins/haiku-enforce/tests"]
 python_files = "test_*.py"
 addopts = "-v --tb=short"
